Checking in quite a few changes here. Firslty, a bunch of files defined methods for GetName. I changed this to having methods initialize a name variable stored in the MPBaseObject class. Secondly there were a few XML input cleanups. I removed all warnings except deprecated warnings in the utils when -Wall is used. I removed the compiling of trash/Parameters and trash/QueryCmds as these are not used. I removed many unused variable, and unused defines in the utilities. Combined the utilities defines files into one file. I added a file MPProblemAccess to get around errors in compilation when certain methods need access to things in MPStrategy, but aren't redily available to them. I am adding one sampler, Negate Sampler to eliminate the need for Collision space samplers. I also cleaned up samplers with elimination of unused variables. I fixed two bugs I found, one in workspaceOBPRM and the other in Gauss. I removed a few unnecessary includes from various files. Lastly, I complletely removed the UAS compile option. This will always be compiled when the code is compiled. I believe this was it, I apologize for this checkin touching almost all files. If this causes some problems let me know and I can help anyone who needs help resolving problems.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@2840 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/LocalPlanner/LocalPlanners.h
@@ -9,7 +9,6 @@
 #define LocalPlanners_h
 
 //////////////////////////////////////////////////////////////////////////////////////////
-#include "OBPRMDef.h"
 #include "CollisionDetection.h" //for CDINFO instance, so we can not use forward declaration.
 #include "ValidityChecker.hpp"
 #include "Cfg.h"    //for vector<Cfg>, so we can not use forward declaration.
@@ -603,8 +602,8 @@ GetPathSegment(Environment *_env, Stat_Class& Stats,
          bool savePath, bool saveFailedPath) {
   bool connected = false;
 
-  if (_weight.GetLP() >= 1 && _weight.GetLP() <= selected.size()) {         
-    for (int lpid = _weight.GetLP(); !connected && lpid <= selected.size(); lpid++) {
+  if (_weight.GetLP() >= 1 && _weight.GetLP() <= (int)selected.size()) {         
+    for (int lpid = _weight.GetLP(); !connected && lpid <= (int)selected.size(); lpid++) {
       // clear possible old storage.  
       _ci->path.erase(_ci->path.begin(), _ci->path.end());
       //the local planner takes care of forward and backward connection
@@ -631,12 +630,12 @@ GetLocalPlanner(unsigned int lpid) {
   LocalPlannerMethod<CFG, WEIGHT>* lp = NULL;
   if (lpid <= 0 || lpid > selected.size()) //out of bounds
     lp = NULL;
-  else if (selected[lpid-1]->GetID() == lpid)
+  else if ((unsigned int)selected[lpid-1]->GetID() == lpid)
     lp = selected[lpid-1];
   else {   //Otherwise search the selected vector for the proper local planner
     typename vector<LocalPlannerMethod<CFG, WEIGHT>*>::iterator itr;
     for (itr = selected.begin(); itr != selected.end(); itr++) {
-      if ( (*itr)->GetID() == lpid ) {
+      if ( (unsigned int)(*itr)->GetID() == lpid ) {
   lp = (*itr);
   break;
       }
